Setting up the view state for blog and create blog
I have created the routes and components for view and create blogs
I can click on the page and it changes the view to the page
I want to display blogs on. And I have a page for create blog.
Next is structuring the forms for creating and viewing blogs.