Fixed the undefined issue
I was able to fix the fact that blog was returning undefined
when I was clicking the create blog button. It was being referred to
by the wrong title in the action command. All of my buttons are now
facing the same error. I believe it to be a controller issue.